Updated agsmlGeneratorGUI
