+++ .gitignore
@@ -1,1 +1,3 @@
 /JavaApplication2/nbproject/private/
+/build/
+/dist/
