--- AGSMLConverter.properties.bak
@@ -1,9 +1,0 @@
-DictionaryFile=P:\\UKCRD1-TI\\Projects\\14\\geotech1\\GEO-3523\\AGSML\\dist\\schema\\DictionaryAgsml_0.0.11.xml
-DictionaryId=AGS4.04
-ProjectId=Project4
-InputSource=C:\\Users\\simon.DORSETRECLAIM\\Downloads\\107971\\E7037A_Burton SWWM_Final Factual Report AGS4.ags
-OutputFolder=C:\\Users\\simon.DORSETRECLAIM\\Downloads\\107971\\E7037A_Burton SWWM_Final Factual Report AGS4.xml
-TempFolder=P:\\UKCRD1-TI\\Projects\\14\\geotech1\\GEO-3523\\AGSML\\dist
-LocalPort=1045
-Server=ukcrd1pc40002
-ServerPort=1045
