+++ nbproject/project.properties
@@ -3,7 +3,7 @@ annotation.processing.enabled.in.editor=false
 annotation.processing.processors.list=
 annotation.processing.run.all.processors=true
 annotation.processing.source.output=${build.generated.sources.dir}/ap-source-output
-application.title=AGSMLConverter
+application.title=agsmlConverter
 application.vendor=thomsonsj
 auxiliary.org-netbeans-spi-editor-hints-projects.perProjectHintSettingsFile=nbproject/cfg_hints.xml
 build.classes.dir=${build.dir}/classes
@@ -26,7 +26,7 @@ debug.test.classpath=\
 dist.archive.excludes=
 # This directory is removed when the project is cleaned:
 dist.dir=dist
-dist.jar=${dist.dir}/AGSMLConverter.jar
+dist.jar=${dist.dir}/agsmlConverter.jar
 dist.javadoc.dir=${dist.dir}/javadoc
 endorsed.classpath=
 excludes=
@@ -80,7 +80,7 @@ jnlp.signed=false
 jnlp.signing=
 jnlp.signing.alias=
 jnlp.signing.keystore=
-main.class=agsmlGenerator.AGSMLGenerator
+main.class=agsmlGenerator.agsmlConverter
 # Optional override of default Application-Library-Allowable-Codebase attribute identifying the locations where your signed RIA is expected to be found.
 manifest.custom.application.library.allowable.codebase=
 # Optional override of default Caller-Allowable-Codebase attribute identifying the domains from which JavaScript code can make calls to your RIA without security prompts.
@@ -104,5 +104,5 @@ run.test.classpath=\
     ${javac.test.classpath}:\
     ${build.test.classes.dir}
 source.encoding=UTF-8
-src.agsml.dir=agsml
-src.agsmlGenerator.dir=agsmlGenerator
+src.src.dir=src
+test.examples.dir=examples
