+++ nbproject/project.xml
@@ -3,12 +3,13 @@
     <type>org.netbeans.modules.java.j2seproject</type>
     <configuration>
         <data xmlns="http://www.netbeans.org/ns/j2se-project/3">
-            <name>AGSMLConverter</name>
+            <name>agsmlConverter</name>
             <source-roots>
-                <root id="src.agsml.dir"/>
-                <root id="src.agsmlGenerator.dir"/>
+                <root id="src.src.dir"/>
             </source-roots>
-            <test-roots/>
+            <test-roots>
+                <root id="test.examples.dir"/>
+            </test-roots>
         </data>
     </configuration>
 </project>
