+++ src/agsml/AGS_Dictionary.java
