+++ src/agsml/AGS_ReaderText.java
