+++ src/agsml/AGS_XML_Writer.java
