+++ src/agsml/TextAreaHandler.java
