+++ src/agsmlGenerator/JTextAreaOutputStream.java
